From d3466a04345b31dfc62d94fe994b40321a6418ec Mon Sep 17 00:00:00 2001 From: Michael Orlitzky <michael@orlitzky.com> Date: Thu, 29 Sep 2016 15:43:06 -0400 Subject: [PATCH 1/1] Zend/acinclude.m4: don't warn about >=bison-3.0.1  versions.
The PHP configure script will emit a warning if it thinks your bison
version is insufficient:

  WARNING: This bison version is not supported for regeneration of
  the Zend/PHP parsers (found: 3.0, min: 204, excluded: 3.0).

However, there is an error in the test that causes it to treat all
3.0.x versions as if they were 3.0. The result is that users get
warned about a perfectly acceptable version of bison.

This patch is meant only for Gentoo, where we can require a working
version of bison (something newer than 3.0.1). Having done so, this
patch removes the check and the WARNING.

Gentoo-Bug: 593278
PHP-Bug: 69055